requirements.txt
asgiref==3.8.1
colorama==0.4.6
datetime-truncate==1.1.1
Django==4.2
django-admin-charts==1.4.0
django-admin-tools-stats==0.9.0
django-bower==5.2.0
django-jazzmin==3.0.1
django-jsonfield==1.4.1
django-light==0.1.0.post3
django-memoize==2.3.1
django-multiselectfield==0.1.13
django-nvd3==0.10.1
django-qsstats-magic==1.1.0
django-suit==0.2.28
djangorestframework==3.15.2
djcacheutils==3.0.0
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.1
pillow==11.0.0
pluggy==1.5.0
psycopg2==2.9.10
pytest==8.3.3
python-dateutil==2.9.0.post0
python-memcached==1.62
python-nvd3==0.16.0
python-slugify==8.0.4
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
tzdata==2024.2
gunicorn
